Major refactoring of the WOPI core endpoints
